refactor: Footer,Header의 폴더 위치를 알맞게 변경
- Footer,Header는  shared로 변경

Header의 경우 shared가 맞는가에 대한 고민이 있음
현재는 헤더가 게시물 관리 시스템으로 고정되어있는데, 이건 정말 고정된 형태인것인가? 혹은 변경될 수 있는가?
 보통 헤더에 메뉴가 담긴경우, 보여지는 페이지에 따라 UI가 업데이트 되는 경우가 있음 그렇다면 Header는 shared 에 넣어선 안됨 우선은 이런경우가 아니여서 shared에 넣고 변경 여지가 생길 시 분리할 예정
